Update NCMBInstallation NCMBCurrentInstallation
