+++ ncmb_unity/Assets/NCMB/NCMBPushScript.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a4efb339912de4baa91fd57a0fd0f474
+guid: a47ca33d60dce4db6ac52fd71bf55cbf
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
