+++ ncmb_unity/Assets/NCMB/Script/NCMBInstallation.cs
@@ -18,12 +18,8 @@ using System;
 using System.IO;
 using System.Threading;
 using System.Collections.Generic;
-using MiniJSON;
 using NCMB.Internal;
-using System.Linq;
-using UnityEngine;
-using System.Runtime.InteropServices;
-using System.Text;
+
 
 namespace  NCMB
 {
@@ -34,60 +30,12 @@ namespace  NCMB
 	public class NCMBInstallation : NCMBObject
 	{
 
-		private void setDefaultProperty ()
-		{
-			IDictionary<string, object> dic = NCMBManager.installationDefaultProperty;
-			object value;
-			if (dic.TryGetValue ("applicationName", out value)) {
-				ApplicationName = (string)value;
-			}
-			if (dic.TryGetValue ("appVersion", out value)) {
-				AppVersion = (string)value;
-			}
-			if (dic.TryGetValue ("deviceType", out value)) {
-				DeviceType = (string)value;
-			}
-			if (dic.TryGetValue ("timeZone", out value)) {
-				TimeZone = (string)value;
-			}
-
-			#if UNITY_ANDROID && !UNITY_EDITOR
-			this["pushType"] = "fcm";
-			#endif
-
-			SdkVersion = CommonConstant.SDK_VERSION;
-		}
-
 		/// <summary>
 		/// コンストラクター。<br/>
 		/// installationsの作成を行います。
 		/// </summary>
-		public NCMBInstallation () : this ("")
+		public NCMBInstallation () : base ()
 		{
-		}
-
-		/// <summary>
-		/// コンストラクター。<br/>
-		/// JSONデータをセットしinstallationを作成する場合、こちらを使用します。
-		/// </summary>
-		internal NCMBInstallation (string jsonText): base () //NCMBObjectのコンストラクタ実行
-		{
-			if (jsonText != null && jsonText != "") {
-				Dictionary<string, object> dic = Json.Deserialize (jsonText) as Dictionary<string, object>;	//辞書形式へ変換
-				object value;
-				if (dic.TryGetValue ("data", out value)) {
-					//iOSのみv1からアップデートした場合は{"data":{"objectId”:”xxxx…
-					dic = (Dictionary<string, object>)value;
-				}
-				//各プロパティの設定
-				_mergeFromServer (dic, false);
-			}
-			//固定値のため、internal化したsetter
-			DeviceToken = NCMBManager._token;
-			//applicationName,appVersion,deviceType,timeZone,SdkVersionを取得/設定
-			#if !UNITY_EDITOR
-			setDefaultProperty ();
-			#endif
 		}
 
 		/// <summary>
@@ -160,28 +108,6 @@ namespace  NCMB
 		}
 
 		/// <summary>
-		/// 現在の配信端末情報を取得します。
-		/// </summary>
-		/// <returns>配信端末情報 </returns>
-		public static NCMBInstallation getCurrentInstallation ()
-		{
-			NCMBInstallation currentInstallation = null;
-			try {
-				//ローカルファイルに配信端末情報があれば取得、なければ新規作成
-				string currentInstallationData = NCMBManager.GetCurrentInstallation ();
-				if (currentInstallationData != "") {
-					//ローカルファイルから端末情報を取得
-					currentInstallation = new NCMBInstallation (currentInstallationData);
-				} else {
-					currentInstallation = new NCMBInstallation ();
-				}
-			} catch (SystemException error) {
-				throw new NCMBException (error);
-			}
-			return currentInstallation;
-		}
-
-		/// <summary>
 		/// installation内のオブジェクトで使用出来るクエリを取得します。
 		/// </summary>
 		/// <returns> クエリ</returns>
@@ -193,46 +119,6 @@ namespace  NCMB
 		internal override string _getBaseUrl ()
 		{
 			return NCMBSettings.DomainURL + "/" + NCMBSettings.APIVersion + "/installations";
-		}
-
-		//SaveAsync通信後の処理
-		internal override void _afterSave (int statusCode, NCMBException error)
-		{
-			if (error != null) {
-				if (error.ErrorCode == "E404001") {
-					//No data available時にcurrentInstallationを削除
-					NCMBManager.DeleteCurrentInstallation (NCMBManager.SearchPath ());
-				}
-			} else if (statusCode == 201 || statusCode == 200) {
-				string path = NCMBManager.SearchPath ();
-				if (path != NCMBSettings.currentInstallationPath) {
-					//成功時にv1のファイルを削除
-					NCMBManager.DeleteCurrentInstallation (path);
-				}
-				_saveInstallationToDisk ("currentInstallation");
-			}
-		}
-
-		internal void _saveInstallationToDisk (string fileName)
-		{
-
-			string path = NCMBSettings.filePath;
-			string filePath = path + "/" + fileName;
-			object obj;
-			Monitor.Enter (obj = this.mutex);
-			try {
-				string jsonData = _toJsonDataForDataFile ();
-
-				//save to file
-				using (StreamWriter sw = new StreamWriter (@filePath, false, Encoding.UTF8)) {
-					sw.Write (jsonData);
-					sw.Close ();
-				}
-			} catch (Exception e) {
-				throw new NCMBException (e);
-			} finally {
-				Monitor.Exit (obj);
-			}
 		}
 
 	}
